Fri, Nov 4, 2022, 10:53 PM +00:00

+++ readability/readability.c
@@ -12,14 +12,14 @@ int main(void)
 {
     string text = get_string("type in text: ");
 
-     int letters = count_letters(text);
-     printf("letters: %i\n", letters);
+     float letters = count_letters(text);
+     printf("letters: %f\n", letters);
 
-    int words = count_words(text);
-    printf("words: %i\n", words);
+    float words = count_words(text);
+    printf("words: %f\n", words);
 
-    int sentenses = count_sentences(text);
-    printf("sentenses: %i\n", sentenses);
+    float sentenses = count_sentences(text);
+    printf("sentenses: %f\n", sentenses);
 
 //Coleman-Liau index variables PROBLEM!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
     float L = (letters/words) * 100;
